// server_web/src/types/resource_type.ts
export interface ResourceItem {
  id?: number;
  resource_parent_id?: number;
  user_id?: number;
  resource_name?: string;
  resource_type?: string;
  resource_type_cn?: string | null;
  resource_title?: string;
  resource_desc?: string;
  resource_icon?: string;
  resource_format?: string;
  resource_size_in_MB?: number;
  resource_path?: string | null;
  resource_status?: string;
  resource_source?: string | null;
  resource_source_url?: string | null;
  resource_show_url?: string | null;
  resource_download_url?: string | null;
  resource_feature_code?: string | null;
  rag_status?: string | null;
  resource_language?: string | null;
  create_time?: string;
  update_time?: string;
  delete_time?: string | null;
  show_buttons?: boolean | null;
  resource_is_selected?: boolean | null;
  resource_parent_name?: string | null;
  sub_resource_dir_cnt?: number | null;
  sub_resource_file_cnt?: number | null;
  sub_rag_file_cnt?: number | null;
  resource_is_supported?: boolean | null;
  resource_view_support?: boolean | null;
  resource_content?: string | null;
  ref_text?: string | null;
  rerank_score?: number | null;
  resource_tags?: ResourceTag[] | null;
  author_info?: Users | null;
  access_list?: string[];
  [property: string]: any;
}
// ...
export interface ResourceTag {
  id: number | null;
  user_id?: number | null;
  tag_name: string | null;
  tag_value: string | null;
  tag_type: string | null;
  tag_source: string | null;
  tag_desc: string | null;
  tag_color: string | null;
  tag_icon: string | null;
  tag_status?: string | null;
  create_time?: string;
  update_time?: string;
  tag_count?: number | null;
  tag_active?: boolean | null;
}
// ...
export interface ISearchByKeywordParams {
  resource_keyword: string;
  resource_id?: string;
  rag_enhanced?: string;
  auth_type?: string;
}